Largest prime factor of a large number

+++ more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	int n = 612852475143;
+	long int n = 612852475143;
 	int i = 2;
 
 	int div n / i;
